+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * add `id` option that allow to create same terminal using hot reload [#978](https://github.com/jcubic/jquery.terminal/issues/978)
 * allow using Object URLs in links [#982](https://github.com/jcubic/jquery.terminal/issues/982)
 * experimental `--cols` CSS variable [#956](https://github.com/jcubic/jquery.terminal/issues/956)
+* add `terminal::lines()` [#966](https://github.com/jcubic/jquery.terminal/issues/966)
 ### Bugfix
 * fix `terminal::login()` when user already authenticated [#980](https://github.com/jcubic/jquery.terminal/issues/980)
 * improve mobile support